test(tui): split rinkaku-tui/src/lib.rs tests by topic
Move the 1260-line inline `#[cfg(test)] mod tests` block out of
`rinkaku-tui/src/lib.rs` into `rinkaku-tui/src/lib_tests/`, one file
per crate-private function under test, following the ADR 0028
`#[path = "..."] mod tests;` pattern already used by `app/mod.rs`.

Topics: `translate_key`, `translate_mouse`, `recompute_and_reload`,
`hunk_jump`, `scroll_clamp`, `goto_dispatch`. Shared fixtures
(`empty_report`, `report_with_one_symbol`, `report_with_symbols_and_edges`,
`dummy_view`, `candidate`) live in `lib_tests/mod.rs` as
`pub(super) fn`.

`lib.rs` drops from 2260 to 1002 lines. Test count and assertions
unchanged (524 rinkaku_tui tests still pass).
